Protección de rutas y configuración para fotos propias

+++ src/controllers/photo.controller.js
@@ -2,7 +2,8 @@ const Photo = require('../models/Photo.model');
 
 // Guardar una foto
 const savePhoto = async (req, res) => {
-  const { photo } = req.body; // Foto en base64 desde el frontend
+  const { photo } = req.body; // Solo recibe la foto del cuerpo de la solicitud
+  const userId = req.user.id; // Obtén el userId del usuario autenticado
 
   if (!photo) {
     return res.status(400).json({ message: 'No se proporcionó una foto.' });
@@ -15,7 +16,7 @@ const savePhoto = async (req, res) => {
   }
 
   try {
-    const newPhoto = new Photo({ image: photo });
+    const newPhoto = new Photo({ image: photo, userId });
     await newPhoto.save();
     res.status(201).json({ message: 'Foto guardada exitosamente', data: newPhoto });
   } catch (error) {
@@ -26,8 +27,10 @@ const savePhoto = async (req, res) => {
 
 // Obtener todas las fotos
 const getAllPhotos = async (req, res) => {
+  const userId = req.user.id; // Obtener el ID del usuario autenticado
+
   try {
-    const photos = await Photo.find();
+    const photos = await Photo.find({ userId }); // Filtrar por userId
     res.status(200).json(photos);
   } catch (error) {
     console.error('Error al obtener las fotos:', error);
@@ -38,12 +41,20 @@ const getAllPhotos = async (req, res) => {
 // Eliminar una foto
 const deletePhoto = async (req, res) => {
   const { photoId } = req.params;
+  const userId = req.user._id; // Obtener el ID del usuario autenticado
 
   try {
-    const deletedPhoto = await Photo.findByIdAndDelete(photoId);
-    if (!deletedPhoto) {
+    const photo = await Photo.findById(photoId);
+    if (!photo) {
       return res.status(404).json({ message: 'Foto no encontrada' });
     }
+
+    // Verificar que el usuario es el propietario de la foto
+    if (photo.userId.toString() !== userId.toString()) {
+      return res.status(403).json({ message: 'No tienes permiso para eliminar esta foto' });
+    }
+
+    const deletedPhoto = await Photo.findByIdAndDelete(photoId);
     res.status(200).json({ message: 'Foto eliminada exitosamente' });
   } catch (error) {
     console.error('Error al eliminar la foto:', error);
